Dynamic routing config and dummy data.
Added a dummy data class that returns a list of predicted values from generated random data. It uses linear least squares to perform this prediction. This class is subject to change based on the development of the front end and scrapper.

Added a config file tht the server reads on instantiation that contains all of the technologies and companies the API can perform predictions on.